google_books/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==2.0.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
